further improved architecture
