--- test/newmain_test.dart
@@ -1,18 +1,0 @@
-import 'package:carbpro/newmain.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:mockito/annotations.dart';
-
-// flutter pub run build_runner build --delete-conflicting-outputs
-
-// @GenerateMocks([CarbPro])
-void main(){
-  group('GOURP ???', () {
-    test('Test???', () {
-
-    });
-
-
-
-  });
-
-}
